@@ -1,6 +1,6 @@
-import events from '@devo/applications-builder/libs/events';
-import scroll from '@devo/applications-builder/utils/scroll';
-import DataNode from '@devo/applications-builder/data/DataNode';
+import events from '@devoinc/applications-builder/libs/events';
+import scroll from '@devoinc/applications-builder/utils/scroll';
+import DataNode from '@devoinc/applications-builder/data/DataNode';
 
 /**
  * @class
