@@ -1,10 +1,10 @@
-import processStructure from '@devo/applications-data-library/structures/series';
-import objects from '@devo/applications-builder/utils/objects';
+import processStructure from '@devoinc/applications-data-library/structures/series';
+import objects from '@devoinc/applications-builder/utils/objects';
 import highchartsHelper from './helpers/highchartsHelper';
-import downloads from '@devo/applications-builder/libs/downloads';
-import HCP from '@devo/applications-builder/widgets/helpers/exportDataSeries';
-import { __ } from '@devo/applications-builder/i18n';
-import widgetFactory from '@devo/applications-builder/widgetFactory';
+import downloads from '@devoinc/applications-builder/libs/downloads';
+import HCP from '@devoinc/applications-builder/widgets/helpers/exportDataSeries';
+import { __ } from '@devoinc/applications-builder/i18n';
+import widgetFactory from '@devoinc/applications-builder/widgetFactory';
 import dependencies from '../data/dependencies';
 
 const LineWidget = dependencies.require('widgets').LineWidget;
