@@ -1,10 +1,10 @@
 import '@babel/polyfill';
 import sinon from 'sinon';
 import template from './template';
-import DataNode from '@devo/applications-builder/data/DataNode';
+import DataNode from '@devoinc/applications-builder/data/DataNode';
 import { SIGINT } from 'constants';
-jest.mock('@devo/applications-builder/utils/scroll');
-import scroll from '@devo/applications-builder/utils/scroll';
+jest.mock('@devoinc/applications-builder/utils/scroll');
+import scroll from '@devoinc/applications-builder/utils/scroll';
 
 beforeEach(() => {
   let $ = (window.jQuery = window.$ = require('jquery'));
@@ -20,13 +20,13 @@ afterEach(() => {
 
 describe('Global Tab Tests: ', () => {
   test('should exist', () => {
-    const Tab = require('@devo/applications-builder/Tab');
+    const Tab = require('@devoinc/applications-builder/Tab');
     expect(Tab).toBeDefined();
   });
 
   test(`Should create a new tab with default settings when is called only
   with an HTML id selector`, () => {
-    const Tab = require('@devo/applications-builder/Tab').default;
+    const Tab = require('@devoinc/applications-builder/Tab').default;
     let tab = new Tab('main1');
     let data = new DataNode('main1');
     expect(Array.isArray(tab.sections)).toBeTruthy();
@@ -36,7 +36,7 @@ describe('Global Tab Tests: ', () => {
   });
 
   test(`Should throw an error when no HTML id is provided`, () => {
-    const Tab = require('@devo/applications-builder/Tab').default;
+    const Tab = require('@devoinc/applications-builder/Tab').default;
     let tab = new Tab('main17');
     tab.init();
     sinon.assert.calledWith(stub, `Tab "main17" does not exists on HTML.`);
@@ -45,7 +45,7 @@ describe('Global Tab Tests: ', () => {
   test(`Should initialize the tab when init is called`, () => {
     const initSubmenuSpy = sinon.spy();
     const addListenersSpy = sinon.spy();
-    const Tab = require('@devo/applications-builder/Tab').default;
+    const Tab = require('@devoinc/applications-builder/Tab').default;
     const initSubMenu = Tab.prototype._initSubMenu;
     const addListeners = Tab.prototype._addListeners;
     Tab.prototype._initSubMenu = initSubmenuSpy;
@@ -71,8 +71,8 @@ describe('Global Tab Tests: ', () => {
 
   test(`Should init child sections when init is called`, () => {
     const initSectionSpy = sinon.spy();
-    const Tab = require('@devo/applications-builder/Tab').default;
-    const Section = require('@devo/applications-builder/Section').default;
+    const Tab = require('@devoinc/applications-builder/Tab').default;
+    const Section = require('@devoinc/applications-builder/Section').default;
     Section.prototype.init = initSectionSpy;
     let tab = new Tab('main1');
     tab.addSection(new Section('section11'));
@@ -89,10 +89,10 @@ describe('Global Tab Tests: ', () => {
   test(`Should add event listeners to tab and its sections when
 addListeners is called`, () => {
     const selectSpy = sinon.spy();
-    const Tab = require('@devo/applications-builder/Tab').default;
+    const Tab = require('@devoinc/applications-builder/Tab').default;
     const select = Tab.prototype.select;
     Tab.prototype.select = selectSpy;
-    const Section = require('@devo/applications-builder/Section').default;
+    const Section = require('@devoinc/applications-builder/Section').default;
     let tab = new Tab('main1');
     tab.addSection(new Section('section11'));
     tab.init();
@@ -106,7 +106,7 @@ addListeners is called`, () => {
   });
 
   test(`Should init submenus`, () => {
-    const Tab = require('@devo/applications-builder/Tab').default;
+    const Tab = require('@devoinc/applications-builder/Tab').default;
     let tab = new Tab('main5');
     tab.init();
     let selector = tab.selector;
@@ -118,7 +118,7 @@ addListeners is called`, () => {
   });
 
   test(`Should set transition options when setTransitions is called`, () => {
-    const Tab = require('@devo/applications-builder/Tab').default;
+    const Tab = require('@devoinc/applications-builder/Tab').default;
     let tab = new Tab('main1');
     tab.setTransition('test');
     expect(tab.trans).toBe('test');
@@ -128,7 +128,7 @@ addListeners is called`, () => {
     const closeSubmenuSpy = sinon.spy();
     const scrollToElementStub = sinon.stub(scroll, 'scrollToElement');
 
-    const Tab = require('@devo/applications-builder/Tab').default;
+    const Tab = require('@devoinc/applications-builder/Tab').default;
     const closeSubmenu = Tab.prototype.closeSubmenu;
     Tab.prototype.closeSubmenu = closeSubmenuSpy;
     let tab = new Tab('main1');
@@ -141,7 +141,7 @@ addListeners is called`, () => {
   });
 
   test(`Should close submenu when closeSubmenu is called`, () => {
-    const Tab = require('@devo/applications-builder/Tab').default;
+    const Tab = require('@devoinc/applications-builder/Tab').default;
     let tab = new Tab('main1');
     tab.init();
     tab.selector.classList.add('scroll-menu-open');
@@ -152,8 +152,8 @@ addListeners is called`, () => {
   });
 
   test(`Should refresh the tab and its sections when tab is active and refresh is called`, () => {
-    const Tab = require('@devo/applications-builder/Tab').default;
-    const Section = require('@devo/applications-builder/Section').default;
+    const Tab = require('@devoinc/applications-builder/Tab').default;
+    const Section = require('@devoinc/applications-builder/Section').default;
     const sectionRefreshSpy = sinon.spy();
     Section.prototype.refresh = sectionRefreshSpy;
     let tab = new Tab('main1');
@@ -167,8 +167,8 @@ addListeners is called`, () => {
   });
 
   test(`Should do nothing when refresh is called and the tab is not active`, () => {
-    const Tab = require('@devo/applications-builder/Tab').default;
-    const Section = require('@devo/applications-builder/Section').default;
+    const Tab = require('@devoinc/applications-builder/Tab').default;
+    const Section = require('@devoinc/applications-builder/Section').default;
     const sectionRefreshSpy = sinon.spy();
     Section.prototype.refresh = sectionRefreshSpy;
     let tab = new Tab('main1');
@@ -183,11 +183,11 @@ addListeners is called`, () => {
   });
 
   test(`Should hide sections when unselect is called`, () => {
-    const Tab = require('@devo/applications-builder/Tab').default;
+    const Tab = require('@devoinc/applications-builder/Tab').default;
     const closeSubmenu = Tab.prototype.closeSubmenu;
     const closeSubmenuSpy = sinon.spy();
     Tab.prototype.closeSubmenu = closeSubmenuSpy;
-    const Section = require('@devo/applications-builder/Section').default;
+    const Section = require('@devoinc/applications-builder/Section').default;
     const sectionhideSpy = sinon.spy();
     Section.prototype.hide = sectionhideSpy;
     let tab = new Tab('main1');
@@ -206,7 +206,7 @@ addListeners is called`, () => {
   });
 
   test(`Should init submenus if the selector has submenu`, () => {
-    const Tab = require('@devo/applications-builder/Tab').default;
+    const Tab = require('@devoinc/applications-builder/Tab').default;
     let tab = new Tab('main5');
     tab.init();
     tab._initSubMenu();
@@ -214,7 +214,7 @@ addListeners is called`, () => {
   });
 
   test(`Should not init submenus if the selector  does not have submenu`, () => {
-    const Tab = require('@devo/applications-builder/Tab').default;
+    const Tab = require('@devoinc/applications-builder/Tab').default;
     let tab = new Tab('main1');
     tab.init();
     tab._initSubMenu();
